backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==5.0.0
python-multipart==0.0.6
aiofiles==23.2.1
slowapi==0.1.9
redis==5.0.1
sqlalchemy==2.0.23
PyPDF2==3.0.1
python-docx==1.2.0
openai==1.3.7
tiktoken==0.5.2
cryptography>=41.0.0
bleach==6.1.0
lxml==4.9.3
regex==2023.10.3
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
